+++ src/ecs/resource/particle.rs
@@ -2,7 +2,7 @@
 
 use {
   std::{
-    f64::consts::{ PI, },
+    f32::consts::{ PI, },
   },
   crate::{
     util::{
@@ -13,25 +13,25 @@ use {
 
 #[derive(Default, Debug)]
 pub struct ParticleRules {
-  pub base_pos: Vector3f<f64>,
-  pub base_dir: Vector3f<f64>,
+  pub base_pos: Vector3f<f32>,
+  pub base_dir: Vector3f<f32>,
   pub texture: String,
   pub tex_rows: u32,
-  pub pps: f64,
-  pub dir_error: f64,
-  pub half_life: f64,
-  pub life_error: f64,
-  pub half_speed: f64,
-  pub speed_error: f64,
-  pub half_scale: f64,
-  pub scale_error: f64,
-  pub angle: f64,
-  pub gravity_mult: f64,
+  pub pps: f32,
+  pub dir_error: f32,
+  pub half_life: f32,
+  pub life_error: f32,
+  pub half_speed: f32,
+  pub speed_error: f32,
+  pub half_scale: f32,
+  pub scale_error: f32,
+  pub angle: f32,
+  pub gravity_mult: f32,
   pub rand_rot: bool,
   pub is_directional: bool,
 }
 impl ParticleRules {
-  pub fn set_direction(&mut self, direction: Vector3f<f64>, deviation: f64) {
+  pub fn set_direction(&mut self, direction: Vector3f<f32>, deviation: f32) {
     self.base_dir = direction;
     self.dir_error = deviation * PI;
     self.is_directional = true;
